--- src/main/java/domain/game/GameState.java
@@ -1,5 +1,0 @@
-package domain.game;
-
-public enum GameState {
-    PLAY, QUIT
-}
