+++ src/main/java/janggi/unit/HorseUnitRule.java
@@ -1,8 +1,8 @@
-package domain.unit;
+package janggi.unit;
 
-import domain.position.Point;
-import domain.position.Position;
-import domain.position.Route;
+import janggi.position.Point;
+import janggi.position.Position;
+import janggi.position.Route;
 import java.util.ArrayList;
 import java.util.List;
