+++ src/test/java/janggi/unit/CarUnitRuleTest.java
@@ -1,21 +1,21 @@
-package domain.unit;
+package janggi.unit;
 
-import domain.position.Position;
-import domain.position.Route;
+import janggi.position.Position;
+import janggi.position.Route;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class BombUnitRuleTest {
+class CarUnitRuleTest {
     @Test
-    @DisplayName("포는 수평/수직으로만 움직일 수 있고 한칸 이상 이동해야 한다")
+    @DisplayName("차는 수평/수직으로만 움직일 수 있다")
     void test1() {
         // given
-        BombUnitRule bombUnitRule = new BombUnitRule();
+        CarUnitRule carUnitRule = new CarUnitRule();
 
         // when
-        List<Route> routes = bombUnitRule.calculateAllRoute(new Position(0, 0));
+        List<Route> routes = carUnitRule.calculateAllRoute(new Position(0, 0));
 
         // then
         Assertions.assertThat(routes).isNotEmpty();
@@ -37,6 +37,7 @@ class BombUnitRuleTest {
                 Route.of(List.of(
                         new Position(1, 0), new Position(2, 0), new Position(3, 0))),
                 Route.of(List.of(new Position(1, 0), new Position(2, 0))),
+                Route.of(List.of(new Position(1, 0))),
                 Route.of(List.of(
                         new Position(0, 1), new Position(0, 2), new Position(0, 3), new Position(0, 4),
                         new Position(0, 5), new Position(0, 6), new Position(0, 7), new Position(0, 8),
@@ -58,6 +59,7 @@ class BombUnitRuleTest {
                 Route.of(List.of(
                         new Position(0, 1), new Position(0, 2), new Position(0, 3))),
                 Route.of(List.of(
-                        new Position(0, 1), new Position(0, 2))));
+                        new Position(0, 1), new Position(0, 2))),
+                Route.of(List.of(new Position(0, 1))));
     }
 }
